requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0
django-autocomplete-light==3.11.0
django-import-export==3.3.7
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.11.0
unicodecsv==0.14.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
